requirements.txt
nextcord
ruamel.yaml
python-dotenv
pymongo